requirements.txt
flask
psycopg2-binary
bcrypt
flask-jwt-extended
flask-smorest
python-dotenv
gunicorn